breaking: [N4S][VEST] rewrite build config
