--- config/builds/index.js
@@ -1,5 +1,0 @@
-import n4s from './n4s';
-import vest from './vest';
-import vestUtils from './vest.utils';
-
-export default [].concat(n4s, vest, vestUtils);
