--- config/builds/main.js.tmpl
@@ -1,8 +1,0 @@
-
-'use strict'
-
-if (process.env.NODE_ENV === 'production') {
-  module.exports = require('./${name}.${format}.min.production.js')
-} else {
-  module.exports = require('./${name}.${format}.development.js')
-}
