+++ config/index.js
@@ -2,7 +2,7 @@ const { readdirSync } = require('fs');
 const path = require('path');
 const { PACKAGES_DIR } = require('../shared/constants.js');
 
-const ROOT_PATH = process.cwd();
+const ROOT_PATH = path.resolve(__dirname, '..');
 const CONFIG_PATH = path.resolve(ROOT_PATH, 'config');
 const PACKAGES_PATH = path.resolve(ROOT_PATH, PACKAGES_DIR);
 const BABEL_CONFIG_PATH = path.resolve(CONFIG_PATH, 'babel', 'babel.config.js');
