+++ config/rollup/nameByFormat.js
@@ -1,0 +1,9 @@
+const FORMAT_ES = 'es';
+
+module.exports = function nameByFormat(format) {
+  if (format === FORMAT_ES) {
+    return 'mjs';
+  }
+
+  return format;
+};
