+++ package.json
@@ -19,12 +19,15 @@
   "jest": {
     "testEnvironment": "node"
   },
+  "private": true,
+  "workspaces": [
+    "packages/*"
+  ],
   "devDependencies": {
     "@ampproject/rollup-plugin-closure-compiler": "^0.26.0",
     "@babel/cli": "^7.12.1",
     "@babel/core": "^7.12.3",
     "@babel/plugin-proposal-optional-chaining": "^7.12.1",
-    "@babel/plugin-proposal-pipeline-operator": "^7.12.1",
     "@babel/plugin-transform-object-assign": "^7.12.1",
     "@babel/plugin-transform-runtime": "^7.12.1",
     "@babel/preset-env": "^7.12.1",
@@ -34,11 +37,9 @@
     "@rollup/plugin-commonjs": "^15.1.0",
     "@rollup/plugin-node-resolve": "^9.0.0",
     "@rollup/plugin-replace": "^2.3.3",
-    "anyone": "^0.0.7",
     "babel-eslint": "^10.1.0",
     "babel-jest": "^26.6.0",
     "babel-plugin-add-module-exports": "^1.0.4",
-    "context": "^1.1.0",
     "cross-env": "^7.0.2",
     "date-fns": "^2.16.1",
     "eslint": "^7.11.0",
@@ -52,17 +53,11 @@
     "jest": "^26.6.0",
     "jest-cli": "^26.6.0",
     "lodash": "^4.17.20",
-    "mocha": "^8.2.0",
     "node-fetch": "^2.6.1",
     "onchange": "^7.1.0",
-    "paraphrase": "^1.8.0",
     "prettier": "2.1.2",
     "pretty-quick": "^3.1.0",
     "rollup": "^2.32.1",
-    "rollup-plugin-babel": "^4.3.2",
-    "rollup-plugin-commonjs": "^10.1.0",
-    "rollup-plugin-node-resolve": "^5.2.0",
-    "rollup-plugin-replace": "^2.2.0",
     "rollup-plugin-terser": "^7.0.2",
     "semver": "^7.3.2",
     "typescript": "^4.0.3",
