+++ packages/n4s/config/rollup/rollup.config.js
@@ -1,0 +1,5 @@
+import enforce from './enforce';
+import enforceExtended from './enforceExtended';
+import ensure from './ensure';
+
+export default [].concat(enforce, ensure, enforceExtended);
