+++ packages/n4s/package.json
@@ -3,6 +3,39 @@
   "version": "2.1.1",
   "description": "Enforce - Validation assertions library",
   "main": "./n4s.js",
+  "browser": "./n4s.js",
+  "module": "./esm/n4s.mjs.production.js",
+  "exports": {
+    ".": {
+      "browser": "./n4s.js",
+      "import": "./esm/n4s.mjs.production.js",
+      "require": "./n4s.cjs.production.js",
+      "node": "./n4s.cjs.production.js",
+      "default": "./n4s.cjs.production.js"
+    },
+    "./esm/n4s.mjs.development.js": "./esm/n4s.mjs.development.js",
+    "./esm/n4s.mjs.production.js": "./esm/n4s.mjs.production.js",
+    "./esm/n4s.mjs.production.min.js": "./esm/n4s.mjs.production.min.js",
+    "./n4s.cjs.development.js": "./n4s.cjs.development.js",
+    "./n4s.cjs.production.js": "./n4s.cjs.production.js",
+    "./n4s.cjs.production.min.js": "./n4s.cjs.production.min.js",
+    "./n4s.umd.development.js": "./n4s.umd.development.js",
+    "./n4s.js": "./n4s.js",
+    "./n4s.umd.production.js": "./n4s.umd.production.js",
+    "./n4s.umd.production.min.js": "./n4s.umd.production.min.js",
+    "./esm/ensure.mjs.development.js": "./esm/ensure.mjs.development.js",
+    "./esm/ensure.mjs.production.js": "./esm/ensure.mjs.production.js",
+    "./esm/ensure.mjs.production.min.js": "./esm/ensure.mjs.production.min.js",
+    "./ensure.js": "./ensure.js",
+    "./ensure.cjs.development.js": "./ensure.cjs.development.js",
+    "./ensure.cjs.production.js": "./ensure.cjs.production.js",
+    "./ensure.cjs.production.min.js": "./ensure.cjs.production.min.js",
+    "./ensure.umd.development.js": "./ensure.umd.development.js",
+    "./ensure.umd.index.js": "./ensure.umd.index.js",
+    "./ensure.umd.production.js": "./ensure.umd.production.js",
+    "./ensure.umd.production.min.js": "./ensure.umd.production.min.js",
+    "./package.json": "./package.json"
+  },
   "repository": {
     "type": "git",
     "url": "git+https://github.com/ealush/vest.git"
@@ -40,5 +73,9 @@
     "rollup-plugin-replace": "^2.2.0",
     "rollup-plugin-terser": "^6.1.0",
     "validator": "^13.0.0"
+  },
+  "scripts": {
+    "build": "rollup -c ./config/rollup/rollup.config.js",
+    "test": "jest"
   }
 }
