+++ packages/vest/config/jest/jest.setup.js
@@ -3,7 +3,7 @@ const { packageJson, packagePath } = require('../../../../util');
 
 const { version } = packageJson(PACKAGE_VEST);
 global.LIBRARY_NAME = PACKAGE_VEST;
-global.VEST_VERSION = version;
+global.__LIB_VERSION__ = version;
 global.ENV_DEVELOPMENT = true;
 
 // Registers global instance
