+++ packages/vest/config/rollup/rollup.config.js
@@ -1,0 +1,4 @@
+import vest from './vest';
+import utilities from './vest.utils';
+
+export default [].concat(vest, utilities);
