+++ packages/vest/src/index.js
@@ -1,9 +1,9 @@
-import enforce from '../../n4s/src/enforce';
+import enforce from 'n4s';
 import create from './core/suite/create';
 import test from './core/test';
 import { draft, only, skip, warn, group } from './hooks';
 
-const VERSION = VEST_VERSION;
+const VERSION = __LIB_VERSION__;
 
 export default {
   VERSION,
