+++ util/index.js
@@ -1,6 +1,7 @@
 const asyncForeach = require('./asyncForeach');
 const exec = require('./exec');
 const logger = require('./logger');
+const packageDist = require('./packageDist');
 const packageJson = require('./packageJson');
 const packagePath = require('./packagePath');
 
@@ -8,6 +9,7 @@ module.exports = {
   asyncForeach,
   exec,
   logger,
+  packageDist,
   packageJson,
   packagePath,
 };
