+++ util/packageDist.js
@@ -1,0 +1,7 @@
+const packagePath = require('./packagePath');
+
+function packageDist(packageName) {
+  return packagePath(packageName, 'dist');
+}
+
+module.exports = packageDist;
